requirements.txt
Pillow
dload
numpy
opencv-python
